Fix mypy (#30)
This is a temporary fix to unblock CI.

+++ src/dvc_objects/fs/system.py
@@ -96,13 +96,16 @@ def _reflink_linux(src: "AnyFSPath", dst: "AnyFSPath") -> None:
 
     FICLONE = 0x40049409
 
-    try:
-        with open(src, "rb") as s, open(dst, "wb+") as d:
-            fcntl.ioctl(d.fileno(), FICLONE, s.fileno())
-    except OSError:
-        with suppress(OSError):
-            os.unlink(dst)
-        raise
+    if sys.platform != "linux":
+        raise AssertionError
+    else:
+        try:
+            with open(src, "rb") as s, open(dst, "wb+") as d:
+                fcntl.ioctl(d.fileno(), FICLONE, s.fileno())
+        except OSError:
+            with suppress(OSError):
+                os.unlink(dst)
+            raise
 
 
 def reflink(source: "AnyFSPath", link_name: "AnyFSPath") -> None:
